feat: cleanup and extract mcu_pwr
